refactor: Move single page measurements to new file
